+++ packages/terra-list/src/terra-dev-site/doc/list/guides.7/SingleSelectList.1.doc.jsx
@@ -1,0 +1,26 @@
+import React from 'react';
+import DocTemplate from 'terra-doc-template';
+import ReadMe from '../../../../../docs/guides/SingleSelectList.md';
+import { name } from '../../../../../package.json';
+
+/* eslint-disable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
+// Example Files
+import SingleSelectList from '../../guides/SingleSelectList';
+import SingleSelectListSrc from '!raw-loader!../../../../../src/terra-dev-site/doc/guides/SingleSelectList';
+/* eslint-enable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
+
+const DocPage = () => (
+  <DocTemplate
+    packageName={name}
+    readme={ReadMe}
+    srcPath={`https://github.com/cerner/terra-core/tree/master/packages/${name}`}
+    examples={[
+      {
+        example: <SingleSelectList />,
+        source: SingleSelectListSrc,
+      },
+    ]}
+  />
+);
+
+export default DocPage;